pyproject.toml
[project]
name = "xbol-documents"
version = "0.1.0"
description = "Documentation for the XBOL project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter-book>=2.1.0",
]

[project.optional-dependencies]
diagrams = [
    "mermaid-cli>=0.1.2",
    "pyyaml>=6.0",
    "watchdog>=4.0",
]

[project.scripts]
render-diagrams = "scripts.render_diagrams:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
